LayoutTests/fast/scrolling/mac/scrollend-event-user-scroll-basic-expected.txt


Test scroll over content
PASS overflowScrollendEventCount is 1
PASS windowScrollendEventCount is 0

Test scroll over document
PASS overflowScrollendEventCount is 0
PASS windowScrollendEventCount is 1
PASS successfullyParsed is true

TEST COMPLETE
